dialects: Add a reshape_ops_utils.py - move ReassociationAttr with its constraint to reshape_ops_utils.py Inspired by the MLIR counterpart: https://github.com/llvm/llvm-project/blob/main/mlir/include/mlir/Dialect/Utils/ReshapeOpsUtils.h
Used by #4245
